+++ backend/core/src/assets/assets.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from "@nestjs/testing"
 import { AssetsController } from "./assets.controller"
 import { AuthModule } from "../auth/auth.module"
-import dbOptions = require("../../ormconfig.test")
+import dbOptions from "../../ormconfig.test"
 import { TypeOrmModule } from "@nestjs/typeorm"
 import { AssetsService } from "./services/assets.service"
 
@@ -9,6 +9,7 @@ describe("AssetsController", () => {
   let controller: AssetsController
 
   beforeEach(async () => {
+    console.log("dbOptions = ", dbOptions)
     const module: TestingModule = await Test.createTestingModule({
       controllers: [AssetsController],
       imports: [TypeOrmModule.forRoot(dbOptions), AuthModule],