+++ backend/core/src/auth/services/user.service.ts
@@ -32,7 +32,6 @@ import { authzActions } from "../enum/authz-actions.enum"
 import { userFilterTypeToFieldMap } from "../dto/user-filter-type-to-field-map"
 import { Application } from "../../applications/entities/application.entity"
 import { Jurisdiction } from "../../jurisdictions/entities/jurisdiction.entity"
-import { UserQueryFilter } from "../filters/user-query-filter"
 import { assignDefined } from "../../shared/utils/assign-defined"
 import { EmailService } from "../../email/email.service"
 import { RequestMfaCodeDto } from "../dto/request-mfa-code.dto"
@@ -41,6 +40,8 @@ import { MfaType } from "../types/mfa-type"
 import { SmsMfaService } from "./sms-mfa.service"
 import { GetMfaInfoDto } from "../dto/get-mfa-info.dto"
 import { GetMfaInfoResponseDto } from "../dto/get-mfa-info-response.dto"
+import { addFilters } from "../../shared/query-filter"
+import { UserFilterParams } from "../dto/user-filter-params"
 
 import advancedFormat from "dayjs/plugin/advancedFormat"
 import { UserRepository } from "../repositories/user-repository"
@@ -94,9 +95,16 @@ export class UserService {
     const qb = this.userRepository.getQb()
 
     if (params.filter) {
-      const filter = new UserQueryFilter()
-      filter.addFilters(params.filter, userFilterTypeToFieldMap, distinctIDQB)
-      filter.addFilters(params.filter, userFilterTypeToFieldMap, qb)
+      addFilters<Array<UserFilterParams>, typeof userFilterTypeToFieldMap>(
+        params.filter,
+        userFilterTypeToFieldMap,
+        distinctIDQB
+      )
+      addFilters<Array<UserFilterParams>, typeof userFilterTypeToFieldMap>(
+        params.filter,
+        userFilterTypeToFieldMap,
+        qb
+      )
     }
 
     if (params.search) {