+++ backend/core/src/listings/listings.module.ts
@@ -10,12 +10,22 @@ import { User } from "../auth/entities/user.entity"
 import { Property } from "../property/entities/property.entity"
 import { TranslationsModule } from "../translations/translations.module"
 import { AmiChart } from "../ami-charts/entities/ami-chart.entity"
+import { ListingFeatures } from "./entities/listing-features.entity"
 import { ActivityLogModule } from "../activity-log/activity-log.module"
 import { ListingRepository } from "./repositories/listing.repository"
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Listing, Preference, Unit, User, Property, AmiChart, ListingRepository]),
+    TypeOrmModule.forFeature([
+      Listing,
+      Preference,
+      Unit,
+      User,
+      Property,
+      AmiChart,
+      ListingRepository,
+      ListingFeatures,
+    ]),
     AuthModule,
     TranslationsModule,
     ActivityLogModule,