+++ backend/core/src/listings/listings.service.ts
@@ -78,6 +78,10 @@ export class ListingsService {
       )
     }
 
+    if (params.search) {
+      innerFilteredQuery.andWhere("listings.name ILIKE :search", { search: `%${params.search}%` })
+    }
+
     // TODO(avaleske): Typescript doesn't realize that the `paginate` bool is a
     // type guard, but it will in version 4.4. Once this codebase is upgraded to
     // v4.4, remove the extra type assertions on `params.limit` below.