+++ sites/partners/cypress/integration/04-application.spec.ts
@@ -9,7 +9,7 @@ describe("Application Management Tests", () => {
 
   it("Application grid should display correct number of results", () => {
     cy.visit("/")
-    cy.get(`[col-id="status"]`).eq(2).contains("Open").click()
+    cy.getByTestId("listing-status-cell").eq(1).click()
     cy.getByID("lbTotalPages").contains("20")
     cy.get(".applications-table")
       .first()