+++ sites/public/pages/applications/household/student.tsx
@@ -66,7 +66,12 @@ const ApplicationHouseholdStudent = () => {
 
   return (
     <FormsLayout>
-      <FormCard header={listing?.name}>
+      <FormCard
+        header={{
+          isVisible: true,
+          title: listing?.name,
+        }}
+      >
         <ProgressNav
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}