+++ sites/public/pages/applications/start/autofill.tsx
@@ -96,7 +96,12 @@ export default () => {
 
   return previousApplication ? (
     <FormsLayout>
-      <FormCard header={listing?.name}>
+      <FormCard
+        header={{
+          isVisible: true,
+          title: listing?.name,
+        }}
+      >
         <ProgressNav
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}