+++ sites/public/pages/applications/start/choose-language.tsx
@@ -94,7 +94,12 @@ const ApplicationChooseLanguage = () => {
 
   return (
     <FormsLayout>
-      <FormCard header={listing?.name}>
+      <FormCard
+        header={{
+          isVisible: true,
+          title: listing?.name,
+        }}
+      >
         <ProgressNav
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}